+++ apps/endringsmelding-pleiepenger/e2e/cypress/e2e/tester/1-utfylling-av-dialog/1.2-flere-arbeidsgivere.cy.ts
@@ -15,8 +15,8 @@ describe('Endre arbeidstid for flere arbeidsgivere', () => {
         cy.visit(cyHelpers.startUrl);
     });
     cyHelpers.startSøknad({ endreArbeidstid: true });
-    cyHelpers.endreEnkeltuke();
-    cyHelpers.endreFlereUker();
+    cyHelpers.endreArbeidEnkeltuke();
+    cyHelpers.endreArbeidFlereUker();
     cyHelpers.fortsettTilOppsummering();
     cyHelpers.kontrollerOppsummering();
     cyHelpers.bekreftOpplysningerOgSendInn();