+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @navikt/sif-ettersending
 
+## 3.17.0
+
+### Minor Changes
+
+-   c976be57: Oppdatere til Aksel 5.1.0 + andre mindre oppdateringer
+
+### Patch Changes
+
+-   Updated dependencies [c976be57]
+    -   @navikt/sif-common-amplitude@2.20.0
+    -   @navikt/sif-common-formik-ds@0.4.0
+    -   @navikt/sif-common-soknad-ds@10.4.0
+    -   @navikt/sif-common-core-ds@9.4.0
+    -   @navikt/sif-common-sentry@0.21.0
+    -   @navikt/sif-common-utils@3.39.0
+
+## 3.16.0
+
+### Minor Changes
+
+-   026fc373: Oppdatere til versjon 5 av designsystemet aksel
+
+### Patch Changes
+
+-   Updated dependencies [41d0a03c]
+-   Updated dependencies [026fc373]
+    -   @navikt/sif-common-formik-ds@0.3.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
+
 ## 3.15.0
 
 ### Minor Changes
@@ -11,8 +41,8 @@
 -   Updated dependencies
     -   @navikt/sif-common-amplitude@2.19.0
     -   @navikt/sif-common-formik-ds@0.24.0
-    -   @navikt/sif-common-soknad-ds@11.0.0
-    -   @navikt/sif-common-core-ds@10.0.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
     -   @navikt/sif-common-sentry@0.20.0
     -   @navikt/sif-common-utils@3.38.0
 
@@ -59,8 +89,8 @@
 
 -   Updated dependencies [ea79f31c]
     -   @navikt/sif-common-formik-ds@0.23.0
-    -   @navikt/sif-common-soknad-ds@11.0.0
-    -   @navikt/sif-common-core-ds@10.0.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
     -   @navikt/sif-common-sentry@0.19.0
 
 ## 3.13.0